+++ src/the_music_tree_genre_kit/criteria/playlist/CriterialessPlaylistNames.py
@@ -1,0 +1,3 @@
+class CriterialessPlaylistNames:
+    GENRE = "Genreless"
+    TAG = "Tagless"