+++ src/the_music_tree_genre_kit/criteria/playlist/Fields.py
@@ -1,0 +1,8 @@
+class Fields:
+    CRITERIA = "criteria"
+    PARENT = "parent"
+    ROOT = "root"
+    NAME = "name"
+    CHILDREN = "children"
+    DESCENDANTS = "descendants"
+    ROOT_DESCENDANTS = "root_descendants"